Remove dynamic require (#119)
* Remove dynamic require

I'm using this module (through npm.im/got) for a serverless function and I'm getting this warning:

```
3:31:45 PM: WARNING in ../node_modules/keyv/src/index.js 18:14-40
3:31:45 PM: Critical dependency: the request of a dependency is an expression
3:31:45 PM:  @ ../node_modules/cacheable-request/src/index.js
3:31:45 PM:  @ ../node_modules/got/dist/source/core/index.js
3:31:45 PM:  @ ../node_modules/got/dist/source/create.js
3:31:45 PM:  @ ../node_modules/got/dist/source/index.js
3:31:45 PM:  @ ./api.js
```

It's not technically a problem, but it is an annoyance and I'd love to avoid the warning.

This simply removes the dynamic require for static and lazy requires. Just a refactor. No functionality change here.

* fix constructor call