Replace `get<Raw>(...)` with an overload
One overload with `options?: { raw?: false }` and one with `options: { raw: true }` allows being specific about the return type for each.

So, the behavior is unchanged but it becomes easier for external libraries to write an interface with a correct *partial* implementation.

My use-case: I'm working on a library that expects a keyv-like object passed in. The user is allowed to pass in a different implementation, as long as the basic `get`, `set` and `delete` exist. But they don't need to provide an implementation for the `{ raw: true }` variant. So, I'm trying to write an interface that corresponds to that, which `new Keyv<string>(...)` conforms to. This is hard/impossible to do without some `as any` hacks right now, but the change in this commit makes it possible.

Signed-off-by: Misha Kaletsky <15040698+mmkal@users.noreply.github.com>